+++ examples/scripts-search-plugin/src/gui.rs
@@ -506,7 +506,7 @@ impl ScriptsSearchApp {
 
     fn handle_non_filter_click(effect: &crate::EffectData, response: &egui::Response) {
         if response.clicked() {
-            let res = crate::EDIT_HANDLE.get().unwrap().call_edit_section(|e| {
+            let res = crate::EDIT_HANDLE.call_edit_section(|e| {
                 let created =
                     e.create_object(&effect.effect.name, e.info.layer, e.info.frame, None)?;
                 e.focus_object(&created)?;
@@ -517,67 +517,66 @@ impl ScriptsSearchApp {
     }
 
     fn add_filter_to_focused_object(effect: &crate::EffectData) -> anyhow::Result<()> {
-        crate::EDIT_HANDLE
-            .get()
-            .unwrap()
-            .call_edit_section(|edit| {
-                // フィルターを追加するAPIがないため、エイリアスを編集して対応する
-                let focused_object = edit
-                    .get_focused_object()?
-                    .ok_or_else(|| anyhow::anyhow!("オブジェクトが選択されていません。"))?;
-                let alias_str = edit.object(&focused_object).get_alias()?;
-                let mut alias: aviutl2::alias::Table = alias_str
-                    .parse()
-                    .map_err(|e| anyhow::anyhow!("Failed to parse alias: {}", e))?;
-                let alias_table = alias
-                    .get_table_mut("Object")
-                    .ok_or_else(|| anyhow::anyhow!("Failed to get Object table from alias"))?;
-                let last_table = alias_table.subtables().last().ok_or_else(|| {
-                    anyhow::anyhow!("Failed to get last subtable from Object table")
-                })?;
-                let effect_index =
-                    last_table.0.parse::<u32>().map_err(|e| {
-                        anyhow::anyhow!("Failed to parse last subtable index: {}", e)
-                    })? + 1;
-                alias_table.insert_table(&effect_index.to_string(), {
-                    let mut table = aviutl2::alias::Table::new();
-                    table.insert_value("effect.name", &effect.effect.name);
-                    table
-                });
-                let base_position = edit.object(&focused_object).get_layer_frame()?;
-                edit.object(&focused_object).delete_object()?;
-                match edit.create_object_from_alias(
-                    &alias.to_string(),
-                    base_position.layer,
-                    base_position.start,
-                    0,
-                ) {
-                    Ok(created) => {
-                        edit.focus_object(&created)?;
-                        anyhow::Ok(())
-                    }
-                    Err(err) => {
-                        edit.create_object_from_alias(
-                            &alias_str,
-                            base_position.layer,
-                            base_position.start,
-                            0,
-                        )?;
-                        Err(err.into())
-                    }
+        crate::EDIT_HANDLE.call_edit_section(|edit| {
+            // フィルターを追加するAPIがないため、エイリアスを編集して対応する
+            let focused_object = edit
+                .get_focused_object()?
+                .ok_or_else(|| anyhow::anyhow!("オブジェクトが選択されていません。"))?;
+            let alias_str = edit.object(&focused_object).get_alias()?;
+            let mut alias: aviutl2::alias::Table = alias_str
+                .parse()
+                .map_err(|e| anyhow::anyhow!("Failed to parse alias: {}", e))?;
+            let alias_table = alias
+                .get_table_mut("Object")
+                .ok_or_else(|| anyhow::anyhow!("Failed to get Object table from alias"))?;
+            let last_table = alias_table
+                .subtables()
+                .last()
+                .ok_or_else(|| anyhow::anyhow!("Failed to get last subtable from Object table"))?;
+            let effect_index = last_table
+                .0
+                .parse::<u32>()
+                .map_err(|e| anyhow::anyhow!("Failed to parse last subtable index: {}", e))?
+                + 1;
+            alias_table.insert_table(&effect_index.to_string(), {
+                let mut table = aviutl2::alias::Table::new();
+                table.insert_value("effect.name", &effect.effect.name);
+                table
+            });
+            let base_position = edit.object(&focused_object).get_layer_frame()?;
+            edit.object(&focused_object).delete_object()?;
+            match edit.create_object_from_alias(
+                &alias.to_string(),
+                base_position.layer,
+                base_position.start,
+                0,
+            ) {
+                Ok(created) => {
+                    edit.focus_object(&created)?;
+                    anyhow::Ok(())
                 }
-            })?
+                Err(err) => {
+                    edit.create_object_from_alias(
+                        &alias_str,
+                        base_position.layer,
+                        base_position.start,
+                        0,
+                    )?;
+                    Err(err.into())
+                }
+            }
+        })?
     }
 
     fn add_filter_as_object(effect: &crate::EffectData) -> anyhow::Result<()> {
-        crate::EDIT_HANDLE.get().unwrap().call_edit_section(|e| {
+        crate::EDIT_HANDLE.call_edit_section(|e| {
             e.create_object(&effect.effect.name, e.info.layer, e.info.frame, None)?;
             anyhow::Ok(())
         })?
     }
 
     fn add_filter_as_filter_object(effect: &crate::EffectData) -> anyhow::Result<()> {
-        crate::EDIT_HANDLE.get().unwrap().call_edit_section(|e| {
+        crate::EDIT_HANDLE.call_edit_section(|e| {
             let filter =
                 e.create_object("フィルタオブジェクト", e.info.layer, e.info.frame, None)?;
             let mut filter_alias = e.object(&filter).get_alias_parsed()?;